unified-backend/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
pydantic-settings==2.7.1
python-dotenv==1.0.1
qdrant-client==1.13.2
numpy
langchain-text-splitters==0.3.5
PyPDF2==3.0.1
boto3==1.37.1
langchain_ollama==0.2.3
langchain==0.3.17
pdfplumber==0.11.5
pytz==2025.1
pymupdf==1.25.2
pandas==2.2.3
fuzzywuzzy==0.18.0
openai
deep-translator==1.11.4
aioboto3==14.1.0
motor==3.7.0
botocore==1.37.1
ultralyticsplus
google-api-python-client==2.148.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-auth==2.35.0
python-multipart